# api/openapi.yaml
openapi: 3.1.0
info:
  title: Verbum API
  version: 0.1.0
  description: |
    The contract between the Verbum apps (iOS, Android) and the backend (docs/PRODUCT.md §45).
    The apps' client interfaces — `GraphClient`, `SearchClient`, `ContextClient`, `TimelineClient`,
    `DailyVerses` — were written first; this document is their wire shape, and `examples/` holds
    the responses both apps decode in their contract tests. Change the JSON here, and the app tests
    break before a user does.

    Conventions
    - Editorial GET routes are public and cacheable. Paid POST routes require a Firebase ID token.
    - Public search is lexical; a verified identity enables paid semantic retrieval.
    - Scripture text is NOT served here: the apps read it from bible.helloao.org (and the bundled WEB).
      References are structured (`PassageReference`), never prose, so the apps render them in the
      reader's translation and language.
    - Every relationship, detail and dating claim carries `sourceReferenceIds` (§33). Content only
      reaches this API through the editorial pipeline with human review (§32).
    - Errors are one shape (`Problem`) with a small `code` vocabulary (§52); the apps map codes to
      states and never show `message` to users.
    - Ids are opaque strings. Fixture ids look like `fixture.person.david`; production ids may differ.
security: []   # editorial content stays public; paid operations override this below
servers:
  - url: https://api.vendlydigital.com.br
...
  - url: http://localhost:8080
    description: local backend
...
paths:
...
  /v1/tts/config:
    get:
      tags: [tts]
      operationId: speechConfiguration
      summary: Default speech version, without contacting the speech provider.
      description: Cached for 3600 seconds. Changes with default voice/settings or assembly revision.
      parameters:
        - name: language
          in: query
          required: true
          schema: { type: string, enum: [pt-BR, en-US] }
      responses:
        '200':
          description: Current audio version.
          content:
            application/json:
              schema:
                type: object
                required: [version]
                properties:
                  version: { type: string, pattern: '^[a-f0-9]{64}$' }
        '400':
          description: Unsupported or missing language.
  /v1/tts:
    post:
      security:
        - FirebaseIDToken: []
      tags: [tts]
      operationId: synthesizeSpeech
      summary: Generate MP3 reading audio with Google Cloud Text-to-Speech.
      description: |
        Verifies the supplied text against the canonical chapter identified by bookId, chapter and translation.
        Arbitrary text, unsupported voices, nonzero pitch and provider-side speed changes are rejected.
        Shared cached audio is replayed before checking generation quotas; only cache misses reserve budget.
        Maximum 100000 UTF-8 bytes of text and 1 MiB request body. Plain text only.
        Automatically segments at 2200 bytes (below Google's 5000-byte limit) and returns one complete MP3,
        crossfading segment joins to avoid an audible seam, cached persistently by exact chapter text and
        normalized voice/audio settings.
        Defaults to pt-BR-Chirp3-HD-Aoede for Portuguese and en-US-Standard-A for English.
        Speed defaults to 1.0; pitch must be omitted or zero for Chirp 3 HD.
        Generation can take up to 10 minutes; no response audio is sent until complete.
        Voice availability and support for speed/pitch depend on Google; invalid combinations
        return 400. Server-side Google credentials are never returned to clients.
        Responses use Cache-Control: no-store; the server keeps its own persistent chapter cache
        keyed by exact text and normalized settings, invisible to HTTP intermediaries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeechRequest'
            examples:
              portuguese:
                externalValue: examples/tts/request-pt-BR.json
              english:
                externalValue: examples/tts/request-en-US.json
      responses:
        '403':
          description: This plan cannot start the operation.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Idempotency key was already bound to different content.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '200':
          description: MP3 bytes, ready to save or play (not JSON/base64).
          headers:
            X-Verbum-Audio-Cues:
              description: Optional JSON array of measured excerpts with verseStart, verseEnd, start and end (seconds), bound to these MP3 bytes. Present for requests with verses; absent means playback without synchronization.
              schema: { type: string }
          content:
            audio/mpeg:
              schema: { type: string, format: binary }
        '400':
          description: Invalid text, language, voice, speed, pitch, format or JSON.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '408':
          description: Request cancelled.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '415':
          description: Content-Type must be application/json.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429':
          description: Generation quota, pending duplicate, API rate limit or provider quota reached.
          headers:
            Retry-After:
              description: Seconds to wait; supplied for API limits.
              schema: { type: integer, example: 60 }
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '502':
          description: Invalid provider response; tts_failed.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: TTS or Firebase authentication unavailable; tts_unavailable or auth_unavailable.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '504':
          description: Provider timeout; tts_timeout.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /v1/entities/{id}:
    get:
      tags: [entities]
      operationId: getEntityDetail
      summary: One entity's page (§9) — the facts, key passages, sources.
      description: Backs `GraphClient.detail(id)` and `GraphClient.entity(id)` (the apps read `.entity` from it).
      parameters:
        - $ref: '#/components/parameters/Lang'
        - $ref: '#/components/parameters/EntityId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityDetail' }
              examples:
                david: { externalValue: examples/entities/david.json }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/entities:
    get:
      tags: [entities]
      operationId: listEntities
      summary: Every entity of one kind, for the Explore lists (§7). Passage nodes are never listed.
      description: Backs `GraphClient.entities(type)`. Sorted by name in the requested language.
      parameters:
        - name: type
          in: query
          required: true
          schema: { $ref: '#/components/schemas/BibleEntityType' }
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [entities]
                properties:
                  entities:
                    type: array
                    items: { $ref: '#/components/schemas/BibleEntity' }
              examples:
                people: { externalValue: examples/entities/people.json }
  /v1/entities/{id}/graph:
    get:
      tags: [graph]
      operationId: getEntityGraph
      summary: The one-hop neighbourhood (§8, §44). Never the whole graph.
      description: |
        Backs `GraphClient.neighbors(id, limit)`. Undirected: an edge stored as David → Goliath is
        returned for Goliath too. `nodes` are the neighbours (the root is not repeated in it);
        `edges` are every stored relationship among `root + nodes`. The apps ask `limit=24` and show
        a balanced dozen themselves (`GraphLayout.balanced`), so the server should return nodes in a
        stable, meaningful order (by edge confidence desc, then name) rather than by kind.
      parameters:
        - $ref: '#/components/parameters/Lang'
        - $ref: '#/components/parameters/EntityId'
        - name: depth
          in: query
          schema: { type: integer, minimum: 1, maximum: 1, default: 1 }
          description: Only 1 is served today; the parameter exists so the URL matches §44.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 48, default: 12 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GraphSnapshot' }
              examples:
                david: { externalValue: examples/graph/david.json }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/passages/{reference}/context:
    get:
      tags: [context]
      operationId: getPassageContext
      summary: Chapter-level context (§10) — entities, related passages, sources.
      description: |
        Backs `ContextClient.chapter(reference)`. `reference` is `Book.Chapter` in OSIS ids
        (`1Sam.17`, `Ps.23`). Context is per chapter. Optional `verse` adds up to 24
        references to other occurrences of the same STEP entities attested in that verse.
        These are sourced entity associations, not asserted quotations or thematic parallels.
        **Missing coverage is 404 with code `content_unavailable`**, which the apps show as
        "no context yet" — never a generated explanation (§3.5).
      parameters:
        - $ref: '#/components/parameters/Lang'
        - name: reference
          in: path
          required: true
          schema: { type: string, pattern: '^[1-3]?[A-Za-z]+\.[0-9]+$' }
          example: 1Sam.17
        - name: verse
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 176 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PassageContext' }
              examples:
                sam17: { externalValue: examples/context/1Sam.17.json }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/timeline:
    get:
      tags: [timeline]
      operationId: getTimeline
      summary: Every period and event, chronological (§4.2), with dating precision.
      description: |
        Backs `TimelineClient.events()`; with `entity`, backs `TimelineClient.eventsFor(id)`.
        Order: by `startYear`, then longer spans first, unknown dates last. `entityNames` saves the
        apps one round-trip per chip (today they call /entities/{id} for each).
      parameters:
        - $ref: '#/components/parameters/Lang'
        - name: entity
          in: query
          required: false
          schema: { type: string }
          description: Only events whose `entityIds` contain this id.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [events, entityNames]
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/TimelineEvent' }
                  entityNames:
                    type: object
                    additionalProperties: { type: string }
                    description: Names for every id that appears in any `entityIds`.
              examples:
                all: { externalValue: examples/timeline/all.json }
  /v1/search:
    get:
      security:
        - {}
        - FirebaseIDToken: []
      tags: [search]
      operationId: search
      summary: References, books and entities for a query (§27–28).
      description: |
        Backs `SearchClient.search(query, language)`. The apps already parse references and match
        book names on device; the server may repeat that (cheap, and keeps one ranking) but MUST
        serve the entity groups. A direct reference match wins outright (§28): when `passages` is
        non-empty, `books` is empty.

        `passages` (2026-09-12): hybrid lexical (Postgres full-text) + semantic (pgvector cosine
        similarity against a server-side query embedding) retrieval over a Scripture text corpus
        (currently WEB, English only). Ranking blends both channels' scores — a first heuristic,
        not a calibrated relevance model. Verified direct references bypass embeddings on the server.
        A missing identity uses lexical/entity/reference search without a paid embedding call.
        A valid Firebase token (including anonymous users) enables semantic retrieval.
        Rate limits return 429 with Retry-After; invalid supplied tokens return 401.
        Missing Firebase configuration with a supplied token returns 503.
        Search results vary by Authorization. Embedding failures degrade to lexical retrieval.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - $ref: '#/components/parameters/Lang'
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: Firebase verification is not configured for a supplied identity.
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResponse' }
              examples:
                david: { externalValue: examples/search/david.json }
  /v1/daily-verse:
    get:
      tags: [daily]
      operationId: getDailyVerse
      summary: The verse of the day — a reference only; the text comes from the reader's translation.
      description: |
        Not in §45; added for the Home card and the 7:00 notification. The apps compute this on
        device today (`DailyVerses`, deterministic per day) and will keep doing so as a fallback.
        The server answers for a date so the apps can plan 14 mornings ahead in one call.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
          description: First day, inclusive. Default today (server's UTC date).
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 31, default: 1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [verses]
                properties:
                  verses:
                    type: array
                    items:
                      type: object
                      required: [date, reference]
                      properties:
                        date: { type: string, format: date }
                        reference: { $ref: '#/components/schemas/PassageReference' }
              examples:
                week: { externalValue: examples/daily-verse/week.json }
  /v1/ask:
    post:
      security:
        - FirebaseIDToken: []
      tags: [ask]
      operationId: askScripture
      summary: Task 12 — Ask Scripture. Retrieval-grounded synthesis with a hard citation gate (§29-31).
      description: |
        The model never names a Bible reference itself; it only picks from passages this server
        already retrieved via the same hybrid search as `/v1/search` and verified exist. An
        answer that cites nothing verifiable is never shown as if it were grounded — the
        response falls back to `confidence: "low"` with an empty `answer` instead (§73: no
        generation before reliable retrieval). English only (WEB) until a licensed pt-BR
        Scripture text exists (§34). Never streamed — full retrieval always precedes synthesis.
        503 when no OpenAI key is configured; there is no offline/fixture fallback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              additionalProperties: false
              properties:
                question: { type: string, minLength: 1, maxLength: 500 }
                reference:
                  $ref: '#/components/schemas/PassageReference'
                  description: Optional reader anchor. Must select 1–6 verses; corpus text is verified before synthesis.
      responses:
        '403':
          description: This plan cannot start the operation.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Idempotency key was already bound to different content.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '415':
          description: Content-Type must be application/json.
        '429': { $ref: '#/components/responses/RateLimited' }
        '400':
          description: Invalid question or JSON body (one object, maximum 8 KiB).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '502':
          description: Upstream request failed.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AskResponse' }
        '503':
          description: Provider or Firebase authentication is not configured (ask_unavailable, realtime_unavailable or auth_unavailable, as applicable).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example: { code: ask_unavailable, message: "ask is not configured on this server" }
  /v1/realtime/session:
    post:
      security:
        - FirebaseIDToken: []
      tags: [realtime]
      operationId: createRealtimeSession
      summary: A one-use ticket for a metered voice connection through Verbum.
      description: |
        Returns a one-use Verbum ticket valid for 60 seconds. Connect to relayPath on the same
        API host using WebSocket and Authorization Bearer clientSecret. No provider credential
        is exposed. Audio passes through the backend without being persisted. Duration,
        output tokens, response count and spending are enforced there. Responses are no-store.
        Guest sessions cannot use voice. A verbum.limit event ends a restricted conversation.
      parameters:
        - name: model
          in: query
          required: false
          schema: { type: string }
          description: Only omitted, empty, or gpt-realtime is accepted by the API access layer.
          # Clients cannot select an arbitrary paid model.
      responses:
        '403':
          description: This plan cannot start the operation.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Idempotency key was already bound to different content.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '400':
          description: Invalid request or unsupported Realtime model.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '502':
          description: Upstream request failed.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RealtimeSession' }
        '503':
          description: Provider or Firebase authentication is not configured (ask_unavailable, realtime_unavailable or auth_unavailable, as applicable).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example: { code: realtime_unavailable, message: "realtime is not configured on this server" }
  /v1/me/usage:
    get:
      security: [{ FirebaseIDToken: [] }]
      operationId: getUsage
      summary: Current server-owned plan and remaining daily generations.
      responses:
        '200':
          description: UTC reset time and generation ceilings; budgets may restrict access sooner. Never cached.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageStatus' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: Usage storage unavailable; expensive operations remain blocked.
  /v1/realtime/connect:
    get:
      security: [{ VerbumVoiceTicket: [] }]
      operationId: connectVoice
      summary: WebSocket relay with server-enforced duration and spending limits.
      responses:
        '101':
          description: JSON Realtime events; verbum.limit carries code and optional retryAt before close.
        '401': { description: Missing, expired or already consumed one-use ticket. }
        '429': { description: Concurrent conversation, quota or budget limit. }
        '503': { description: Usage storage or provider unavailable. }
components:
  securitySchemes:
    VerbumVoiceTicket:
      type: http
      scheme: bearer
      description: One-use ticket returned by createRealtimeSession, never a provider key.
    FirebaseIDToken:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase SDK ID token for a registered or anonymous user, validated against the configured project including revocation.
  parameters:
    EntityId:
      name: id
      in: path
      required: true
      schema: { type: string }
      example: fixture.person.david
    Lang:
      name: lang
      in: query
      required: false
      schema: { type: string, enum: [en, pt, pt-BR], default: en }
      description: Presentation language; pt aliases pt-BR. Missing translations fall back to en. Original scripts and identifiers remain unchanged. Unsupported values use en.
  responses:
    Unauthenticated:
      description: Missing, invalid, expired, revoked or disabled Firebase identity. No paid operation is started.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Per-user, IP, process or concurrency limit reached. No paid operation is started.
      headers:
        Retry-After:
          schema: { type: integer, example: 60 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Unknown id, or no content for it yet.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { code: unknown_entity, message: "no entity with id 'nope'" }
  schemas:
...
    SpeechRequest:
      type: object
      additionalProperties: false
      required: [text, language, bookId, chapter, translation]
      properties:
        bookId: { type: string, description: Canonical OSIS book identifier, example: John }
        chapter: { type: integer, minimum: 1, maximum: 150 }
        translation: { type: string, enum: [por_blj, BSB], description: Must match language and the canonical text. }
        verses:
          type: array
          minItems: 1
          maxItems: 176
          description: Optional ordered verses for synchronized narration. Their text joined by newlines must equal text exactly. Measured excerpts contain at most three verses (long verses can span several excerpts). Omitting verses returns audio without timing metadata; canonical chapter identifiers remain required.
          items:
            type: object
            additionalProperties: false
            required: [number, text]
            properties:
              number: { type: integer, minimum: 1, maximum: 176 }
              text: { type: string, minLength: 1 }
        revision:
          type: string
          description: Optional version from GET /v1/tts/config. A changed version returns 400 before generation; omitted preserves legacy behavior.
        text:
          type: string
          minLength: 1
          description: Exact chapter plain text, nonblank, at most 100000 UTF-8 bytes. Automatically segmented without truncation.
        language: { type: string, enum: [pt-BR, en-US] }
        voice:
          type: string
          maxLength: 128
          description: Google voice matching language; omitted/empty selects pt-BR-Chirp3-HD-Aoede or en-US-Standard-A.
          example: pt-BR-Chirp3-HD-Aoede
        speed: { type: number, enum: [1], default: 1, description: Adjust playback speed locally without regenerating speech. }
        pitch: { type: number, enum: [0], default: 0 }
        format: { type: string, enum: [MP3], default: MP3 }
    # ---- Errors (§52). Apps switch on `code`; `message` is for logs.
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [unauthenticated, auth_unavailable, rate_limited, unknown_entity, unknown_book, content_unavailable, malformed_request, internal, realtime_unavailable, ask_unavailable, tts_unavailable, tts_rate_limited, tts_timeout, tts_failed, quota_exceeded, budget_exhausted, plan_required, usage_unavailable, request_in_progress, idempotency_conflict]
        message: { type: string }
        retryAt: { type: string, format: date-time }
        fallback: { type: string, enum: [scripture_search] }

    # ---- §22.1
    BibleEntityType:
      type: string
      enum: [person, place, event, theme, passage, book, prophecy, originalTerm, historicalPeriod]
    BibleEntity:
      type: object
      required: [id, type, name]
      properties:
        id: { type: string }
        type: { $ref: '#/components/schemas/BibleEntityType' }
        name: { type: string }
        summary: { type: [string, 'null'], description: Short and non-interpretive (§3.5). }
        nameLanguage:
          type: string
          enum: [en, pt-BR]
          description: Selected presentation record, including fallback. OriginalTerm names keep the original script; its language is lexical.language. Absent for legacy entities without localization records.

    # ---- §22.2
...
    PassageReference:
      type: object
      required: [bookId, chapter]
      properties:
        bookId: { type: string, example: 1Sam }
        chapter: { type: integer, minimum: 1 }
        verseStart: { type: integer, minimum: 1, description: Omitted with verseEnd for a whole chapter. }
        verseEnd: { type: integer, minimum: 1 }

    # ---- §33
...
    OriginalTermPresentation:
      type: object
      required: [language, transliteration, strong]
      properties:
        language: { type: string, enum: [he, arc, grc] }
        transliteration: { type: string }
        strong: { type: string }
    EntityDetail:
      type: object
      required: [entity, aliases, keyPassages, sources]
      properties:
        originalTerm: { $ref: "#/components/schemas/OriginalTermPresentation", description: Canonical study metadata; localized meaning is entity.summary. }
        sourceRecords: { type: array, items: { $ref: '#/components/schemas/EntitySourceRecord' }, description: Optional attributed structured enrichment. }
        localizations: { type: array, items: { $ref: '#/components/schemas/EntityLocalization' }, description: Independent presentation records; absent when unavailable. }
        entity: { $ref: '#/components/schemas/BibleEntity' }
        aliases: { type: array, items: { type: string } }
        approximateDates: { type: [string, 'null'], description: 'Always hedged: "c. 1010–970 BC (commonly dated)".' }
        role: { type: [string, 'null'] }
        modernGeography: { type: [string, 'null'], description: Places only. }
        keyPassages: { type: array, items: { $ref: '#/components/schemas/PassageReference' } }
        sources: { type: array, items: { $ref: '#/components/schemas/SourceReference' } }
...
    GraphSnapshot:
      type: object
      required: [root, nodes, edges]
      properties:
        root: { $ref: '#/components/schemas/BibleEntity' }
        nodes: { type: array, items: { $ref: '#/components/schemas/BibleEntity' } }
        edges: { type: array, items: { $ref: '#/components/schemas/BibleRelationship' } }

    # ---- §10 / §22.4
    PassageContext:
      type: object
      required: [reference, entities, relatedPassages, sources]
      properties:
        reference: { $ref: '#/components/schemas/PassageReference', description: Always a whole chapter. }
        entities: { type: array, items: { $ref: '#/components/schemas/BibleEntity' } }
        relatedPassages:
          type: array
          items: { $ref: '#/components/schemas/PassageReference' }
          description: Share a connected entity with the chapter; not asserted to be quotations or parallels.
        sources: { type: array, items: { $ref: '#/components/schemas/SourceReference' } }

    # ---- §22.5 (+ sourceReferenceIds)
...
    TimelineEvent:
      type: object
      required: [id, title, datePrecision, entityIds, sourceReferenceIds]
      properties:
        id: { type: string }
        title: { type: string }
        startYear: { type: [integer, 'null'], description: Negative = BC (-1010 is 1010 BC). }
        endYear: { type: [integer, 'null'], description: Null with a startYear = a point in time. }
        datePrecision: { $ref: '#/components/schemas/TimelineDatePrecision' }
        summary: { type: [string, 'null'] }
        entityIds: { type: array, items: { type: string } }
        sourceReferenceIds: { type: array, items: { type: string } }

    # ---- §27
...